+++ backend/yourwine/src/main/java/com/ssafy/yourwine/repository/GrapeRepository.java
@@ -1,11 +1,14 @@
 package com.ssafy.yourwine.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.ssafy.yourwine.model.entity.Grape;
 
 
-public interface GrapeRepository extends JpaRepository<Grape, Integer>{
+public interface GrapeRepository extends JpaRepository<Grape, Long>{
 
-	Grape findByGrapeId(int grapeId);
+	Grape findByGrapeId(Long grapeId);
+	
 }
